requirements.txt
torch
torchvision
Pillow
requests
diffusers==0.12.1
transformers
